Moved current weather retrieval to it's own function, and added tweet function.

+++ WeatherBot.py
@@ -7,7 +7,6 @@ import os
 from GetWeather_Data import get_weather_data
 from WeatherBot_Auth import authenticate
 
-
 def arg_check():
     """
     Checks to see if enough arguments are passed to the program.
@@ -18,68 +17,65 @@ def arg_check():
         usage()
         sys.exit(1)
 
-def get_current_weather(location, request_type):
+def get_current_weather(request_type, location):
     """
-    return a string which contains current weather
+    Returns a string with current weather information for tweet.
     """
-    arg_check()
-
     keys_file = sys.argv[1]
-    twitter, auth, weather_key = authenticate(keys_file)
+    twitter, weather_key = authenticate(keys_file)
 
     data = get_weather_data(weather_key, request_type, location)
 
-    location = data["current_observation"]["display_location"]["full"]
-    weahter  = data["current_observation"]["weather"]
-    feeltemp = data["current_observation"]['windchill_string']
-    temp     = data["current_observation"]['temperature_string']
-    time     = data["current_observation"]["observation_time"]
-    wind     = data["current_observation"]["wind_string"]
-    humi     = data["current_observation"]["relative_humidity"]
-    url      = data["current_observation"]["icon_url"]
+    location  = data["current_observation"]["display_location"]["full"]
+    weahter   = data["current_observation"]["weather"]
+    windchill = data["current_observation"]['windchill_string']
+    temp      = data["current_observation"]['temperature_string']
+    time      = data["current_observation"]["observation_time"]
+    wind      = data["current_observation"]["wind_string"]
+    humi      = data["current_observation"]["relative_humidity"]
+    icon_url  = data["current_observation"]["icon_url"]
 
-    tweet_output = location + '\n' + time + '\n' + temp + '\n' + "feeling like "\
-        + feeltemp + '\n' + wind + '\n'
-
-    return tweet_output
-
-def main():
-    """
-    update twitter status on timeline.
-    """
-    arg_check()
-
-    keys_file = sys.argv[1]
-    twitter, auth, weather_key = authenticate(keys_file)
-
-    request_type = "conditions"
-    location = '/q/zmw:12180.1.99999'
-    
-    data = get_weather_data(weather_key, request_type, location)
-
-    location = data["current_observation"]["display_location"]["full"]
-    weahter  = data["current_observation"]["weather"]
-    feeltemp = data["current_observation"]['windchill_string']
-    temp     = data["current_observation"]['temperature_string']
-    time     = data["current_observation"]["observation_time"]
-    wind     = data["current_observation"]["wind_string"]
-    humi     = data["current_observation"]["relative_humidity"]
-    url      = data["current_observation"]["icon_url"]
-
-    tweet_output = location + '\n' + time + '\n' + temp + '\n' + "feeling like "\
-        + feeltemp + '\n' + wind + '\n'
-    filename = 'temp.jpg'
-
-    request = requests.get(url, stream=True)
+    icon_file = 'temp.jpg'
+    request = requests.get(icon_url, stream=True)
     if request.status_code == 200:
         with open(filename, 'wb') as image:
             for chunk in request:
                 image.write(chunk)
 
-    print(tweet_output)
+    tweet_text = location + '\n' \
+               + time     + '\n' \
+               + temp     + '\n' \
+               + "Feels like " + windchill + '\n'
+               + wind + '\n'
 
-    twitter.update_with_media(filename=filename,status=tweet_output)
-    os.remove(filename)
+    return twitter, icon_file, tweet_text
+
+def tweet(twitter_api, text, image_file=""):
+    """
+    Tweets from text and image input.
+    """
+    if len(image_file) > 0:
+        twitter_api.update_with_media(filename=icon_file, status=text)
+        os.remove(filename)
+
+    else:
+        twitter_api.update(status=text)
+
+def main():
+    """
+    Main function to run the program.
+    update twitter status on timeline.
+    """
+    arg_check()
+
+    request_type = "conditions"
+    location = '/q/zmw:12180.1.99999'
+
+    twitter, icon_file, tweet_text = get_current_weather(request_type, location)
+
+    print(tweet_text)
+
+    tweet(twitter, tweet_text, icon_file)
 
 
 # ============================================================================ #
